6° commit: Se agrego AdminControlador, ImagenControlador, lista de Usuarios en usuario-gestor.html, td de imagen en expositor, falta solucionar boton de modificar en lista de usuarios - 24/06/24

+++ FullStack-WebServicio/src/main/java/com/FinalEgg/ServiChacras/entidades/Proveedor.java
@@ -2,7 +2,6 @@ package com.FinalEgg.ServiChacras.entidades;
 
 import java.util.List;
 import jakarta.persistence.*;
-
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
@@ -38,5 +37,5 @@ public class Proveedor {
     private Integer cantPedido;
 
     @Column(name = "promedio_puntuacion")
-    private Integer promPuntuacion;
+    private Float promPuntuacion;
 }
